+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.0.0
+
+* Changed: Don't append the loxone topic as a prefix
+* Added: Option to publish the value as raw instead of json payload
+
 ## 2.3.1
 
 * Fixed: Reorder the mqtt message parsing to support number only payloads